\name{SIGIL-package}
\alias{SIGIL-package}
\alias{SIGIL}
\docType{package}
\title{

SIGIL: Statistical Inference -- A Gentle Introduction for (Computational and Corpus) Linguists

}
\description{

This package provides example data sets and support functions used in the course units and exercises
of \emph{Statistical Inference: A Gentle Introduction for Computational Linguists and Similar Creatures}
developed by Marco Baroni and Stefan Evert.

}
\details{
\tabular{ll}{
Package: \tab SIGIL\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2014-04-24\cr
License: \tab GPL-3\cr
}
}
\author{ 

Marco Baroni [\url{http://clic.cimec.unitn.it/marco/}] and
Stefan Evert [\url{http://www.stefan-evert.de/}]

Maintainer: Stefan Evert <\email{stefan.evert@fau.de}>

}
\references{

The official homepage of the SIGIL course and this companion package is \url{http://SIGIL.R-Forge.R-Project.org/}.

}
\keyword{ package }
\seealso{

\bf{TODO:} entry points to package documentation

}
